requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Werkzeug==2.2.3
pymongo==4.4.0
python-dotenv==1.0.0
pytest==7.3.1
